Poll Glance directly for image information
There is really no need for the calls to get image status and
image list and deleting images to go thru Nova. This change allows
Qonos snapshot worker to directly talk to Glance. This also has
the benefits that polling can be done with service admin credentials
thus helping deal with token expirations and getting the image list
can be more tightly filtered with a much smaller list of images
being fetched (i.e. only images created by the scheduled images
service for the specified instance.)

Note that a few tests were deleted. With the change to using
glance the list filtering will happen in on the glance side, not
in the snapshot processor. The deleted tests were effectively
testing that filtering which has been removed.

Fixes to allow http or https for communication

Responding to review comments. Added some associated tests.