Moving hard timeout and retry notifications back to api.
The hard timeout and retry concepts are meant to protect the api
from never-finishing jobs continually being retried. When this
was moved to the snapshot processor it put the responsibility
for this protection into the hands of the very thing which could
be causing the problem. It was done thinking that only the workers
should be sending notifications, but since these are really API
concepts enforcement and notification should fall to the API.
Therefore moving these checks back.