Allow worker to reload configuration changes without restart
It would be useful to be able to change some worker configuration
values without having to restart all the workers and interrupt
ongoing jobs. This commit adds support for reloading the config
files when a SIGHUP is received. Note that there are some values
which will not take affect until restart regardless.

Also updated the worker config sample.