Adjust to oslo namespace changes and fixes tests
* `oslo.i18n` => `oslo_i18n`
* `oslo.serialization` => `oslo_serialization`
* `oslo.config` => `oslo_config`
* `oslo.utils` => `oslo_utils`

Also, fixes test failures due to the usage of 'assert_called_once' with
mock objects. There is no such method as 'assert_called_once'. There is
'assert_called_once_with' which asserts one call to a method along with
it's arguments. If not that, one can use 'call_count' to assert the
number of calls without checking the arguments.