Process retention for hardtimedout and maxretried
Currently, when a worker picks up a job and if it's HARD_TIMEDOUT or MAX_RETRIED,
the job is marked appropriately and put back into the queue instantly.
This doesn't give the worker a chance to poll the image and process retention if
the image went active since the job was last processed.
As a result, we may be retaining a greater number of images than what the
retention indicates.